🚧 Refactor test file paths for error handling middleware and middleware handling
